fix: move the intro-video presenter manifest into core
The platform localized-UI lint treats the 80 presenter names as hardcoded UI
copy. They are JoggAI catalog names rather than translatable copy, and catalog
data already lives in core alongside the illustration and presentation template
items, so the manifest moves there instead of carrying 80 lint exclusions.

Also reformats the CLI help string with the pinned Prettier version.
